NuConta calculator
Made for Brazilian retail investors, this code allows the user to know how much money his Nubank savings account (NuConta) made in a day, a month or a year. The user should input the amount of money in the account and the current value for the CDI, the index that determines yields. 
The text in the code is in Portuguese.
As of 02/01/2020, CDI = 4,5%