Poupança vs CDI calculator
Made for Brazilian retail investors, this code allows the user to know how much money would make in the Poupança and in an CDI, popular fixed income investments with the same level of risk. In addition to it, the user can know the difference between yields brought by both investments. The user should input how much money it holds.
The code is in Portuguese. The comments are in English.
As of 05/01/2020, Selic = 4.5% and TR = 0%